third_party/blink/web_tests/platform/linux-chrome/external/wpt/html-aam/roles-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] el-code
  assert_equals: <code data-testname="el-code" data-expectedrole="code" class="ex">x</code> expected "code" but got "none"
[FAIL] el-dd
  assert_equals: <dd data-testname="el-dd" data-expectedrole="definition" class="ex">x</dd> expected "definition" but got "DescriptionListDetail"
[FAIL] el-details
  assert_equals: <details data-testname="el-details" data-expectedrole="group" class="ex"><summary>x</summary>x</details> expected "group" but got "Details"
[FAIL] el-dt
  assert_equals: <dt data-testname="el-dt" data-expectedrole="term" class="ex">x</dt> expected "term" but got "DescriptionListTerm"
[FAIL] el-em
  assert_equals: <em data-testname="el-em" data-expectedrole="emphasis" class="ex">x</em> expected "emphasis" but got "none"
[FAIL] el-form
  assert_equals: <form data-testname="el-form" data-expectedrole="form" class="ex"><input></form> expected "form" but got "Section"
[FAIL] el-hgroup
  assert_equals: <hgroup data-testname="el-hgroup" data-expectedrole="group" class="ex"><h1>x</h1></hgroup> expected "group" but got "none"
[FAIL] el-option
  assert_equals: <option data-testname="el-option" data-expectedrole="option" class="ex">x</option> expected "option" but got "menuitem"
[FAIL] el-strong
  assert_equals: <strong data-testname="el-strong" data-expectedrole="strong" class="ex">x</strong> expected "strong" but got "none"
Harness: the test ran to completion.
